+++ CHANGELOG.md
@@ -7,8 +7,9 @@
  * passcode is now checked against a hash just like a real password
  * when user exits the experiment compeleted will now be saved as a timestamp rather than a boolean (so we see how long they stayed in the sim)
  * more intelligent query to load pregenerated ideas / comments
- * liking / unliking now done via channel
+ * liking, unlikeing and posting comments now done via channel and socket
  * delayed post scheduling now done on server and communicated via channel
+ * fixed a bug that caused the deletion of other users's likes on liking a comment
 
 
 ### 2.3.3